+++ pages/views/settings/api.php
@@ -26,13 +26,16 @@ require_once(__ROOT__.'/inc/settings.php');
     </div>
 </div>
 <div class="row mb-3">
-    <div class="col-4">
-        <label for="pv_api_key">API Key</label>
-        <div class="col-md-8 password-input-container">
-        	<input name="pv_api_key" type="password" class="form-control password-input" id="pv_api_key" value="<?=$user['API_key']?>" />
-        	<i class="toggle-password fa fa-eye"></i>
-        </div>
-    </div>
+	<div class="col-4">
+		<label for="pv_api_key">API Key</label>
+		<div class="col-md-8 password-input-container">
+			<input name="pv_api_key" type="password" class="form-control password-input bg-secondary" id="pv_api_key" value="<?=$user['API_key']?>" readonly />
+			<i class="toggle-password fa fa-eye"></i>
+		</div>
+		<div class="col-md-4">
+			<button type="button" id="regenerate-api-key" class="btn btn-warning border border-secondary mt-2">Regenerate</button>
+		</div>
+	</div>
     <div class="col-6">
         <h6>Available API calls</h6>
         <table id="endpointsTable" class="table table-striped nowrap" style="width:100%">
@@ -56,6 +59,7 @@ require_once(__ROOT__.'/inc/settings.php');
 
 <script>
 $(document).ready(function() {
+	$.fn.dataTable.ext.errMode = 'none';
 	var	api_key = 'xxxxxx';
 
 	$(".toggle-password").click(function () {
@@ -75,7 +79,6 @@ $(document).ready(function() {
     });
 	
    $('#endpointsTable').DataTable({
-		$.fn.dataTable.ext.errMode = 'none';
 	   	dom: '',
 		processing: true,
         language: {
@@ -118,13 +121,14 @@ $(document).ready(function() {
 			data: {
 				manage: 'api',		
 				api: $("#pv_api").is(':checked'),
-				api_key: $("#pv_api_key").val(),
 			},
 			dataType: 'json',
 			success: function (data) {
 				if(data.success){
 					$('#toast-title').html('<i class="fa-solid fa-circle-check mx-2"></i>' + data.success);
 					$('.toast-header').removeClass().addClass('toast-header alert-success');
+					$('#pv_api_key').val(data.API_key);
+					reload_data();
 				} else if(data.error) {
 					$('#toast-title').html('<i class="fa-solid fa-warning mx-2"></i>' + data.error);
 					$('.toast-header').removeClass().addClass('toast-header alert-danger');
@@ -132,7 +136,36 @@ $(document).ready(function() {
 				$('.toast').toast('show');
 			},
 			error: function (xhr, status, error) {
-				$('#toast-title').html('<i class="fa-solid fa-circle-exclamation mx-2"></i> An ' + status + ' occurred, check server logs for more info. '+ error);
+				$('#toast-title').html('<i class="fa-solid fa-circle-exclamation mx-2"></i>An ' + status + ' occurred, check server logs for more info. '+ error);
+				$('.toast-header').removeClass().addClass('toast-header alert-danger');
+				$('.toast').toast('show');
+			}
+		});
+	});
+
+
+	$('#regenerate-api-key').click(function() {
+		$.ajax({ 
+			url: '/core/core.php', 
+			type: 'POST',
+			data: {
+				'regenerate-api-key': 'true'
+			},
+			dataType: 'json',
+			success: function (data) {
+				if(data.success){
+					$('#toast-title').html('<i class="fa-solid fa-circle-check mx-2"></i>' + data.success);
+					$('.toast-header').removeClass().addClass('toast-header alert-success');
+					$('#pv_api_key').val(data.API_key);
+					reload_data();
+				} else if(data.error) {
+					$('#toast-title').html('<i class="fa-solid fa-warning mx-2"></i>' + data.error);
+					$('.toast-header').removeClass().addClass('toast-header alert-danger');
+				}
+				$('.toast').toast('show');
+			},
+			error: function (xhr, status, error) {
+				$('#toast-title').html('<i class="fa-solid fa-circle-exclamation mx-2"></i>An ' + status + ' occurred, check server logs for more info. '+ error);
 				$('.toast-header').removeClass().addClass('toast-header alert-danger');
 				$('.toast').toast('show');
 			}