+++ releasenotes.md
@@ -1,5 +1,7 @@
 Whats New in v12.6
 --------------------------
+- Improved API security
+- Implemented a unique auto generate 32 bit key for API access
 - Formula making improvements
 - Added meta data (formulas, ingredients,...) count for users in the admin dashboard
 - Validate size upload limit function improvements